Fixed sound board list length
There was an issue where if there were too many files in the sound board folder, the bot was unable to display/list it in the channel chat with the "sb_list" command due to mumble's text size limitations. This has been fixed by splitting the message into multiple parts if the length exceeds 50.